requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0

# Git repository support
GitPython>=3.1.40

# LLM integrations
ollama>=0.1.0

# Code parsing
tree-sitter>=0.20.4
tree-sitter-java>=0.20.3
tree-sitter-c-sharp>=0.20.0
tree-sitter-php>=0.20.2

# Utilities
pathspec>=0.12.0
rich>=13.7.0
click>=8.1.7
requests>=2.31.0

# Web UI
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-file-browser>=0.1.0

# PDF generation
markdown>=3.5.1
weasyprint>=60.1
pygments>=2.17.2

# Template engine
jinja2>=3.1.2
